refactor: move UnitTabBanner, TagFunctional and TagPromotional
