+++ src/components/CurriculumComponents/UnitsTab/UnitsTab.tsx
@@ -14,8 +14,8 @@ import Radio from "@/components/SharedComponents/RadioButtons/Radio";
 import RadioGroup from "@/components/SharedComponents/RadioButtons/RadioGroup";
 import Sidebar from "@/components/Sidebar/Sidebar";
 import UnitModal from "@/components/CurriculumComponents/UnitModal/UnitModal";
-import { TagFunctional } from "@/components/TagFunctional";
-import UnitTabBanner from "@/components/UnitTabBanner";
+import { TagFunctional } from "@/components/SharedComponents/TagFunctional";
+import UnitTabBanner from "@/components/CurriculumComponents/UnitTabBanner";
 import { P, Heading } from "@/components/SharedComponents/Typography";
 import useAnalytics from "@/context/Analytics/useAnalytics";
 import useAnalyticsPageProps from "@/hooks/useAnalyticsPageProps";
